requirements.txt
lunardate
skyfield
Pillow
